[gn][banjo] Fix some warnings in banjo gn build
[22816->1012/23885 ~58] CXX host-x64-linux-clang/obj/system/host/banjo/formatter/banjo-format.main.cpp.o
../system/host/banjo/formatter/main.cpp:75:16: warning: comparison of integers of different signs: 'int' and 'std::__2::vector<std::__2::basic_string<char>, std::__2::allocator<std::__2::basic_string<char> > >::size_type' (aka 'unsigned long') [-Wsign-compare]
    while (pos < args.size() && args[pos] != "--" && args[pos].find("-") == 0) {
           ~~~ ^ ~~~~~~~~~~~
../system/host/banjo/formatter/main.cpp:87:13: warning: comparison of integers of different signs: 'int' and 'std::__2::vector<std::__2::basic_string<char>, std::__2::allocator<std::__2::basic_string<char> > >::size_type' (aka 'unsigned long') [-Wsign-compare]
    if (pos >= args.size()) {
        ~~~ ^  ~~~~~~~~~~~
../system/host/banjo/formatter/main.cpp:95:16: warning: comparison of integers of different signs: 'int' and 'std::__2::vector<std::__2::basic_string<char>, std::__2::allocator<std::__2::basic_string<char> > >::size_type' (aka 'unsigned long') [-Wsign-compare]
    for (; pos < args.size(); pos++) {
           ~~~ ^ ~~~~~~~~~~~
3 warnings generated.

BLD-361 #comment [gn][banjo] Fix some warnings in banjo gn build
ZX-3415 #comment [gn][banjo] Fix some warnings in banjo gn build

Test: CQ

+++ zircon/system/host/banjo/formatter/main.cpp
@@ -70,7 +70,7 @@ int main(int argc, char* argv[]) {
     }
 
     bool in_place = false;
-    int pos = 1;
+    size_t pos = 1;
     // Process options
     while (pos < args.size() && args[pos] != "--" && args[pos].find("-") == 0) {
         if (args[pos] == "-i" || args[pos] == "--in-place") {
